feat: read processor settings from root of s3 dir
Changes where the config is read from to the root of the S3 dir that
is having its messages processed. This enables easier deployments in
multi-deployed scenarios.

Makefile updates to reduce the size of the generated lambda function
substantially from ~28 MB to ~4 MB.

Removed push_messages dependency as the single dynamodb call to make
was more efficient to encode directly as part of the substantial
reduction in lambda function package size.

Closes #8